Moved new populate_datastore_res_views method to dump module.
