+++ ckanapi/cli/dump.py
@@ -8,14 +8,13 @@ import json
 from datetime import datetime
 import os
 
-from ckanapi.errors import (NotFound, NotAuthorized, ValidationError,
+from ckanapi.errors import (CKANAPIError, NotFound, NotAuthorized, ValidationError,
     SearchIndexError)
 from ckanapi.cli import workers
 from ckanapi.cli.utils import completion_stats, compact_json, \
     quiet_int_pipe
 from ckanapi.datapackage import create_datapackage, \
-    populate_datastore_res_fields, \
-    populate_datastore_res_views
+    populate_datastore_res_fields
 
 
 def dump_things(ckan, thing, arguments,
@@ -213,3 +212,25 @@ def _worker_command_line(thing, arguments):
         + b('--resource-views')
         + ['value-here-to-make-docopt-happy']
         )
+
+
+def populate_datastore_res_views(ckan, res):
+    """
+    update resource dict in-place with resource_view_list values
+    in every resource with datastore active using ckan
+    LocalCKAN/RemoteCKAN instance
+    """
+    if not res.get('datastore_active', False):
+        return
+    try:
+        ds = ckan.call_action('resource_view_list', {
+            'id': res['id'],
+            'limit':0})
+    except CKANAPIError:
+        return
+    except NotFound:
+        return  # with localckan we'll get the real CKAN exception not a CKANAPIError subclass
+    if not ds:
+        return # return if the resource views list is empty
+    res['resource_views'] = ds
+
